\name{catrec}
\docType{data}
\alias{catrec}
\title{A Test for Existence of a GDOR}
\description{
  The \code{catrec} data frame provides an example of when a gdor exists.
}
\usage{catrec}
\format{
  A 2 X 2 X ... X 2 (2^7 = 128 cells) contingency table producing an example
    for when the MLE fails to exist in the conventional sense implying existence
    of a GDOR
}
\references{
  Geyer, Charles J. (2009)
  \emph{Likelihood inference in exponential families and directions of
    recession.}
  Electronic Journal of Statistics 3, (2009), 259-289 (electronic)
}
\keyword{datasets}